requirements.txt
pyOpenSSL
requests
urllib3